classes created : 'ModelVM', 'UserVM', 'User', 'Exam', 'Enums'.
ModelVM: this class will be the base class for any view model, it contains methods and property relevents for any ViewModel.
PropertyChanged  - handle updating the GUI when a bound property changes.
OnPropertyChanged - invoke the 'PropertyChanged' event when a property is changed.

UserVM : this View model connect the GUI and an instance of  'User' class.
fields and properties :
'User' hold an instance of the 'User' class.
'Id', 'UserId', 'FirstName', 'LastName', 'Email', 'PhoneNumber', 'Password', 'UserType'. These properties provide access to the corresponding fields in the User object. The get accessor retrieves the value from the User object, while the set accessor updates the User object and refreshes the GUI.
constructors:
A parameterless constructor initializes a new 'User' object.
A constructor that takes a 'User' object as a parameter and sets the 'User' property.

User represent the user information model.
fields :
'Id' (Primary Key): The unique identifier for the user.
'UserId': User's identification number.
'FirstMame', 'LastName',  'Email', 'PhoneNumber', 'Password', 'UserType' : basic user information.
'ExamsNames'- a list of exams names related to the user.
constructors :
1. A default constructor sets default values: empty strings for textual fields, 'NoType' for 'UserType', and a new list for 'ExamsNames'.
2. A constructor initalize all fields exept for the primery key 'Id'.
methods :
GetFullName - asynchronous method that returns the full name of the user.

Enums.cs
enums will contain all the enums requied for this project.
UserType - an enumeration with 3 values :
1. Teacher.
2. Student.
3. NOTYPE